+++ Documentation/conf.py
@@ -127,3 +127,9 @@ napoleon_use_ivar = True  # solves duplicate object description warning
 
 # nbsphinx configuration
 nbsphinx_execute = "never"  # This is currently not working
+
+latex_elements = {
+    'preamble': r'''
+\usepackage{amsmath} 
+''',
+}